requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
fpdf2>=2.7.0
Pillow>=10.0.0
